refactor(log): split logging headers into level / sink / logging
Break the monolithic log/logging.hpp into three headers so consumers
pull in only what they need: log/level.hpp (the level enum + compile-time
threshold macros + name parser), log/sink.hpp (the sink interface +
set_sink/get_sink), and log/logging.hpp (the SIRIUS_LOG_* macros and
their format_and_log helper). The include chain is logging.hpp ->
sink.hpp -> level.hpp, so a file that only logs sees the logging API up
front rather than the sink setup.

The level enumerators are now defined in terms of the SIRIUS_LOG_LEVEL_*
macros, so the two representations cannot drift and the static_assert is
gone. string_to_enum(level) moves to level.cpp.

Part of #1145.
